Replace cluster detauls page with the new version code
