--- assets/js/components/ChecksResults/index.js
@@ -1,26 +1,0 @@
-import ChecksResults from './ChecksResults';
-import ResultsContainer from './ResultsContainer';
-import HostResultsWrapper from './HostResultsWrapper';
-import CheckResult from './CheckResult';
-import {
-  getHosts,
-  getChecks,
-  getCheckHealthByAgent,
-  getCheckResults,
-  getCheckDescription,
-  getCheckRemediation,
-} from './checksUtils';
-
-export {
-  ResultsContainer,
-  HostResultsWrapper,
-  CheckResult,
-  getHosts,
-  getChecks,
-  getCheckHealthByAgent,
-  getCheckResults,
-  getCheckDescription,
-  getCheckRemediation,
-};
-
-export default ChecksResults;
