+++ assets/js/components/ClusterDetails/index.js
@@ -1,10 +1,7 @@
-import ClusterDetails from './ClusterDetails';
+import { ClusterDetails } from './ClusterDetails';
 
 export { ClusterSettings } from './ClusterSettings';
-export { ConnectionSettings } from './ConnectionSettings';
 export { ExecutionIcon } from './ExecutionIcon';
 export { ClusterInfoBox } from './ClusterInfoBox';
 
 export default ClusterDetails;
-export { ClusterDetailsNew } from './ClusterDetailsNew';
-export { ClusterSettingsNew } from './ClusterSettingsNew';
