+++ assets/js/components/ExecutionResults/CheckResult.jsx
