+++ assets/js/components/ExecutionResults/ExecutionIcon.jsx
