+++ assets/js/components/ExecutionResults/checksUtils.js
