+++ assets/js/components/ExecutionResults/checksUtils.test.js
